+++ examples/pool-connection.rs
@@ -1,6 +1,6 @@
 use scatter_gather_core::{
     middleware_specs::{
-        ServerConfig,
+        NodeConfig,
     },
     pool::{
         Pool,
@@ -9,7 +9,8 @@ use scatter_gather_core::{
     },
 };
 use scatter_gather_websockets::WebSocketsMiddleware;
-use scatter_gather::source_specs::{
+mod source_specs;
+use source_specs::{
     binance::BinanceDepthInterceptor,
     bitstamp::BitstampDepthInterceptor,
 };
@@ -26,13 +27,13 @@ async fn main() -> Result<(), Box<dyn std::error::Error + 'static>> {
     let binance_interceptor = BinanceDepthInterceptor::new();
     let bitstamp_interceptor = BitstampDepthInterceptor::new();
 
-    let config_binance: ServerConfig = ServerConfig {
+    let config_binance: NodeConfig = NodeConfig {
         url : String::from("wss://stream.binance.com:9443/ws/ethbtc@depth@100ms"),
         prefix: String::from("wss://"),
         init_handle: None,
         // handler: binance_interceptor
     };
-    let config_bitstamp: ServerConfig = ServerConfig { 
+    let config_bitstamp: NodeConfig = NodeConfig { 
         url: String::from("wss://ws.bitstamp.net"), 
         prefix: String::from("wss://"), 
         init_handle: Some(r#"{"event": "bts:subscribe","data":{"channel": "diff_order_book_ethbtc"}}"#.to_string()),