+++ examples/pool-to-grpc.rs
@@ -1,7 +1,7 @@
 use futures::StreamExt;
 use scatter_gather_core::{
     middleware_specs::{
-        ServerConfig, Interceptor,
+        NodeConfig, Interceptor,
     },
     pool::{
         Pool,
@@ -22,7 +22,8 @@ use scatter_gather_grpc::{
         }
     }
 };
-use scatter_gather::source_specs::{
+mod source_specs;
+use source_specs::{
     Depth,
     Level as DepthLevel,
     Interceptors,
@@ -48,13 +49,13 @@ async fn main() -> Result<(), Box<dyn std::error::Error + 'static>> {
     let binance_interceptor = BinanceDepthInterceptor::new();
     let bitstamp_interceptor = BitstampDepthInterceptor::new();
 
-    let config_binance: ServerConfig = ServerConfig {
+    let config_binance: NodeConfig = NodeConfig {
         url : String::from("wss://stream.binance.com:9443/ws/ethbtc@depth@100ms"),
         prefix: String::from("wss://"),
         init_handle: None,
         // handler: binance_interceptor
     };
-    let config_bitstamp: ServerConfig = ServerConfig { 
+    let config_bitstamp: NodeConfig = NodeConfig { 
         url: String::from("wss://ws.bitstamp.net"), 
         prefix: String::from("wss://"), 
         init_handle: Some(r#"{"event": "bts:subscribe","data":{"channel": "diff_order_book_ethbtc"}}"#.to_string()),
@@ -75,7 +76,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error + 'static>> {
             .map(|result| result.unwrap().into_text().unwrap())
             .map(|text| Interceptors::Bitstamp(BitstampDepthInterceptor::intercept(text)) );
 
-    let grpc_config: ServerConfig = ServerConfig { 
+    let grpc_config: NodeConfig = NodeConfig { 
         url: String::from("[::1]:54001"), 
         prefix: String::from("http://"), 
         init_handle: None,