+++ examples/ws-simple-connection.rs
@@ -1,11 +1,12 @@
 use scatter_gather_core::{
     middleware_specs::{
-        ServerConfig, 
+        NodeConfig, 
         Interceptor,
     },
 };
 use scatter_gather_websockets::WebSocketsMiddleware;
-use scatter_gather::source_specs::{
+mod source_specs;
+use source_specs::{
     Depth,
     bitstamp::BitstampDepthInterceptor,
 };
@@ -14,13 +15,10 @@ use futures::StreamExt;
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error + 'static>> {
 
-    let bitstamp_interceptor = BitstampDepthInterceptor::new();
-
-    let config: ServerConfig = ServerConfig {
+    let config: NodeConfig = NodeConfig {
         url : String::from("wss://ws.bitstamp.net"),
         prefix: String::from(""),
         init_handle: Some(r#"{"event": "bts:subscribe","data":{"channel": "diff_order_book_ethbtc"}}"#.to_string()),
-        // handler: bitstamp_interceptor
     };
     let mut connection = WebSocketsMiddleware::new(config).await;
     connection.send(r#"{"event": "bts:subscribe","data":{"channel": "diff_order_book_ethbtc"}}"#.to_string()).await;
@@ -29,7 +27,6 @@ async fn main() -> Result<(), Box<dyn std::error::Error + 'static>> {
         .map(|result| result.unwrap().into_text().unwrap())
         .map(|msg| BitstampDepthInterceptor::intercept(msg));
     while let Some(data) = new_stream.next().await {
-        // let data: interceptor = connection.config.handler.intercept(a?.into_text()?);
         println!("Summary: {:?}", data.level());
     }
     Ok(())