+++ middleware/scatter-gather-redis/src/lib.rs
@@ -1,25 +1,25 @@
 use ::redis::{Client, aio::MultiplexedConnection};
 use scatter_gather_core::{
     connection::ConnectionHandlerOutEvent,
-    middleware_specs::ServerConfig
+    middleware_specs::NodeConfig
 };
 use tokio::sync::mpsc;
 mod redis;
 
 #[derive(Debug)]
 pub struct RedisMiddleware {
-    pub config: ServerConfig,
+    pub config: NodeConfig,
     pub write: mpsc::Sender<ConnectionHandlerOutEvent<String>>,
     read: mpsc::Receiver<ConnectionHandlerOutEvent<String>>,
     client: MultiplexedConnection
 }
 
 impl RedisMiddleware {
-    pub async fn new(config: ServerConfig) -> RedisMiddleware {
+    pub async fn new(config: NodeConfig) -> RedisMiddleware {
         Self::spin_up(config).await.expect("Couldn't build Middleware.")
     }
 
-    pub async fn spin_up(config: ServerConfig) -> Result<Self, Box<dyn std::error::Error>> {
+    pub async fn spin_up(config: NodeConfig) -> Result<Self, Box<dyn std::error::Error>> {
         let (write, read): (mpsc::Sender<ConnectionHandlerOutEvent<String>>, mpsc::Receiver<ConnectionHandlerOutEvent<String>>) = mpsc::channel(32);
         let client = Client::open(config.url.clone())?;  // "redis://172.17.0.2/"
         let con = client.get_multiplexed_tokio_connection().await?;