+++ models/build.rs
@@ -1,0 +1,11 @@
+fn main() {
+    prost_build::compile_protos(
+        &[
+            "src/keys.proto",
+            // "src/envelope.proto",
+            // "src/peer_record.proto",
+        ],
+        &["src"],
+    )
+    .unwrap();
+}