+++ core/CHANGELOG.md
@@ -9,7 +9,7 @@
 # 0.2.0 [2023-02-21]
 
 - Add Executor and Handler traits.
-- Add Connection type with ConnectionId and ServerConfig.
+- Add Connection type with ConnectionId and NodeConfig.
 - Add PoolConnection bounded to Handler trait.
 - Add PendingConnection wrapper bounded to Handler trait.
 - Add EstablishedConnection wrapper bounded to Handler trait.